pyproject.toml
[tool.poetry]
name = "og-bots"
version = "0.1.0"
description = ""
authors = ["karrless <konstantin.geydarov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
vkbottle = "^4.3.12"
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"
loguru = "^0.7.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
